Add Badge class
This is a helper class that makes object models more clear. If you
have a member function of class A that you only want run from another
class B, without class B having full friend access to the entirety of
class A, then make a function argument of that member function
Badge<B> without a variable name, since it isnt actually used in the
function. See the CPU constructor for an example. A CPU object can only be
constructed by the NESEmulator global object.