Add a fully working CPU without clock cycle support
This brings a host of fixes and additions that form a functional NES
6502 CPU. The stack functions are no longer inline, and I added all memory
access modes, perfected using the nestest rom. All instructions should
be accounted for, and if there are any missing instructions, they
should already be prepared, but just missing opcode appearances in the
opcode switch table. A simple addition there should quickly fix that.